feat: make `injectImports` & `injectRootOptions` work for `.vue` files
